合併功能修改

+++ ISpanShop.MVC/Areas/Admin/Controllers/Orders/RoleController.cs
@@ -6,10 +6,12 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using ISpanShop.Models.EfModels;
+using Microsoft.AspNetCore.Authorization;
 
 namespace ISpanShop.MVC.Areas.Admin.Controllers.Orders
 {
     [Area("Admin")]
+    [Authorize] // 確保只有登入的使用者可以訪問
     public class RoleController : Controller
     {
         private readonly ISpanShopDBContext _context;
@@ -19,13 +21,13 @@ namespace ISpanShop.MVC.Areas.Admin.Controllers.Orders
             _context = context;
         }
 
-        // GET: Roles
+        // GET: Admin/Role/Index
         public async Task<IActionResult> Index()
         {
             return View(await _context.Roles.ToListAsync());
         }
 
-        // GET: Roles/Details/5
+        // GET: Admin/Role/Details/5
         public async Task<IActionResult> Details(int? id)
         {
             if (id == null)
@@ -43,29 +45,31 @@ namespace ISpanShop.MVC.Areas.Admin.Controllers.Orders
             return View(role);
         }
 
-        // GET: Roles/Create
+        // GET: Admin/Role/Create
+        [Authorize(Roles = "SuperAdmin,Admin")] // 只有 SuperAdmin 和 Admin 可以新增
         public IActionResult Create()
         {
             return View();
         }
 
-        // POST: Roles/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
+        // POST: Admin/Role/Create
         [HttpPost]
         [ValidateAntiForgeryToken]
+        [Authorize(Roles = "SuperAdmin,Admin")]
         public async Task<IActionResult> Create([Bind("Id,RoleName,Description")] Role role)
         {
             if (ModelState.IsValid)
             {
                 _context.Add(role);
                 await _context.SaveChangesAsync();
+                TempData["SuccessMessage"] = "角色新增成功！";
                 return RedirectToAction(nameof(Index));
             }
             return View(role);
         }
 
-        // GET: Roles/Edit/5
+        // GET: Admin/Role/Edit/5
+        [Authorize(Roles = "SuperAdmin,Admin")] // 只有 SuperAdmin 和 Admin 可以編輯
         public async Task<IActionResult> Edit(int? id)
         {
             if (id == null)
@@ -81,11 +85,10 @@ namespace ISpanShop.MVC.Areas.Admin.Controllers.Orders
             return View(role);
         }
 
-        // POST: Roles/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
+        // POST: Admin/Role/Edit/5
         [HttpPost]
         [ValidateAntiForgeryToken]
+        [Authorize(Roles = "SuperAdmin,Admin")]
         public async Task<IActionResult> Edit(int id, [Bind("Id,RoleName,Description")] Role role)
         {
             if (id != role.Id)
@@ -99,6 +102,7 @@ namespace ISpanShop.MVC.Areas.Admin.Controllers.Orders
                 {
                     _context.Update(role);
                     await _context.SaveChangesAsync();
+                    TempData["SuccessMessage"] = "角色更新成功！";
                 }
                 catch (DbUpdateConcurrencyException)
                 {
@@ -116,7 +120,8 @@ namespace ISpanShop.MVC.Areas.Admin.Controllers.Orders
             return View(role);
         }
 
-        // GET: Roles/Delete/5
+        // GET: Admin/Role/Delete/5
+        [Authorize(Roles = "SuperAdmin")] // 只有 SuperAdmin 可以刪除
         public async Task<IActionResult> Delete(int? id)
         {
             if (id == null)
@@ -134,18 +139,28 @@ namespace ISpanShop.MVC.Areas.Admin.Controllers.Orders
             return View(role);
         }
 
-        // POST: Roles/Delete/5
+        // POST: Admin/Role/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
+        [Authorize(Roles = "SuperAdmin")]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var role = await _context.Roles.FindAsync(id);
             if (role != null)
             {
+                // 檢查是否有使用者正在使用此角色
+                var usersWithRole = await _context.Users.CountAsync(u => u.RoleId == id);
+                if (usersWithRole > 0)
+                {
+                    TempData["ErrorMessage"] = $"無法刪除此角色，目前有 {usersWithRole} 位使用者正在使用此角色。";
+                    return RedirectToAction(nameof(Index));
+                }
+
                 _context.Roles.Remove(role);
+                await _context.SaveChangesAsync();
+                TempData["SuccessMessage"] = "角色刪除成功！";
             }
 
-            await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
